move classifiers to methods
