--- elfi/classifiers/__init__.py
@@ -1,1 +1,0 @@
-# noqa: D104
